Unify Solution.components values
This edit is required for any of the tests checking the output of `Solution.components`
to pass. When new solutes are added (through Solution.add_solute) or their concentrations
are changed (through Solution.set_amount), the volume is recalculated. This adjusts the
number of moles of solvent using its density. However, the density is stored as a
numpy.float64, so further calculations to determine the number of moles of solvent will
be done using this numpy float. This fix ensures that whenever the number of moles of solvent
are adjusted, the value in Solution.components is a standard Python float.

Regression tests have also been added in test_solution.py.

Signed-off-by: Ugochukwu Nwosu <ugn@sfu.ca>