solution.py docstring formatting
This commit contains a number of edits made to docstrings to improve the
documentation. Examples include formatting superscripts and subscripts using
the :sup: and :sub: roles, respectively, marking up text so as to render the
text as code (either via indentation or back ticks), fixing formulas for clarity,
formatting math, and adding cross-references.

Signed-off-by: Ugochukwu Nwosu <ugn@sfu.ca>